+++ udp-project/client/defines.h
@@ -76,7 +76,7 @@ typedef struct gobackn_packet{
 extern int 	 	timeout;		//Timeout of the go back n protocol
 extern int 	 	server_port;  	//Num port where main thread wait connections
 extern int 	 	window;    		//the dimension of the window(packets in fly)
-extern double   prob_loss;      //probability of loss
+extern int 	    prob_loss;      //probability of loss
 extern int 		adaptive;		//variable for adaptive timer if it isnt 0
